ced-52 cambio de archivos
