+++ controllers/entrada.controller.js
@@ -116,3 +116,21 @@ exports.updateEntrada = async function (req, res) {
 
     }    
 };
+
+
+exports.getDetallesPorId = async function (req, res) {
+    let result = validationResult(req);
+
+    if (result.errors.length > 0) {
+        res.status(400).json({ success: false, error: result });
+    } else {
+        let idEntrada = req.params.idEntrada;
+        let entradaDetalles = await entradaService.detallesPorId(idEntrada);
+
+        if (entradaDetalles !== undefined) {
+            res.json(entradaDetalles).status(200);
+        } else {
+            res.status(204).json({ success: false });
+        }        
+    }
+};
