+++ models/index.js
@@ -27,6 +27,5 @@ db.Usuario = require("./evento.model.js")(sequelize, Sequelize);
 db.Almacen = require("./evento.model.js")(sequelize, Sequelize);
 db.Salida = require("./salida.model.js")(sequelize, Sequelize);
 
-db.EntradaDetalle = require("./entrada_detalles.model.js")(sequelize, Sequelize);
 db.Productos = require("./productos.model.js")(sequelize, Sequelize);
 module.exports = db;
